HW02. Move MealTo, add userId to Meal

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -24,23 +24,38 @@ public class InMemoryMealRepository implements MealRepository {
             repository.put(meal.getId(), meal);
             return meal;
         }
-        // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        boolean check = checkUser(meal.getId(), meal.getUserId());
+        if (check) {
+            return repository.put(meal.getId(), meal);
+        }
+        return null;
     }
 
     @Override
-    public boolean delete(int id) {
-        return repository.remove(id) != null;
+    public boolean delete(int id, int userId) {
+        if (checkUser(id, userId)) {
+            return repository.remove(id) != null;
+        }
+        return false;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        if (checkUser(id, userId)) {
+            return repository.get(id);
+        }
+        return null;
     }
 
     @Override
     public Collection<Meal> getAll() {
         return repository.values();
+    }
+
+    @Override
+    public boolean checkUser(int id, int userId) {
+        return repository.values().stream()
+                .anyMatch(m -> (m.getUserId().equals(userId) && m.getId().equals(id)));
     }
 }
